Update online_train.py
Updated creation of generated_sequences_mask. The previous version will only return True or False (in this case False). The current update will set all values of the mask which are not equal to the pad_token_id as 1 and those equal to pad_token_id as 0. This mask will have the same shape as generated_sequences